requirements.txt
aiogram~=3.2.0
environs~=9.5.0
python-dotenv~=1.0.0
psycopg2-binary
requests
